fix: slight chance of getting stuck on channel open
in case the connection is closed right before trying to re-open a
channel amqp.Connection.Channel() will block. It looks like the Go
channel for the notification of a closed amqp channel is blocked causing
a dead-lock. If we add a buffer of one message the method returns and we
can stop trying to open a channel. We then return to trying to open a
new connection and then come back to open a new channel.

From Channel.NotifyClose
> In case of a non graceful close the error will be notified synchronously by the library
so that it will be necessary to consume the Channel from the caller in order to avoid deadlocks