fix: register consumers again after re-connecting DEVOPS-366 (#72)
* fix: re-consume after re-establishing connection or channel

* keep track of consumers via their consumer id generated on the call to
  Consume
* setup the consumers receiver functions on re-opening a connection or
  channel

chore: fix godoc and improve test

* feat: allow setting connection name

* docs: improve readme and example

* chore: write lock Consume and tracking consumer

so that the state of consumers that successfully went through Consume() and
consumeFunc are guaranteed to be in sync

* test: allow disabling connections to producer/consumer separately

via toxiproxy in the docker compose example.

* fix: exit on the first failed consumer registration

* fix: re-declare queue when re-registering consumers

In the scenario that the consumer is connected to RabbitMQ and RabbitMQ shuts down
all non-durable queues are deleted. Our re-registration logic was only registering the
consumer but not declaring the queue again. This is why we got a 404 trying to consume
from a queue that does not exist

* chore: reuse registerConsumerFunc

as we now are doing exactly the same the first time we Consume
or when retrying. Also acquiring locks is now done outside the func

* chore: lowercase error message

---------

Co-authored-by: Andreas Jensen <andreas@dhis2.org>