chore: log using slog
use the slog.Default() logger by default so you can use
the consumer if you do not care about how logging is configured.
Allow setting the logger with an option. Show how this is done
in the example consumer with a logger from which we create a new
one for the consumer that uses a group

stick to word https://www.rabbitmq.com/docs/consumers#consumer-tags
instead of consumer ID which does not seem to be used.